+++ module/enrichers.js
@@ -1,5 +1,5 @@
 import { TestDialog } from './test-dialog.js';
-import { torgDamage, torgDamageModifiers } from './torgchecks.js';
+import { torgDamage, torgDamageModifiers, checkUnskilled } from './torgchecks.js';
 
 /**
  * INLINE CHECKS
@@ -143,6 +143,8 @@ function _onClickInlineCheck(event) {
     else if (actor.type === 'threat')
       skillValue += Math.max(skill.value, attribute.value);
     const isInteractionAttack = (test.attack || interactionAttacks.includes(skillName));
+
+    if (!test.unskilledUse && checkUnskilled(skill.value, skillName, actor)) return;
 
     foundry.utils.mergeObject(test, {
       testType: isInteractionAttack ? 'interactionAttack' : 'skill',